[MISC] Integration test tweaks (#405)
* Remove skipped legacy tests

* Scale down one by one in the legacy test

* Scale down one by one
