+++ tests/integration/new_relations/test_new_relations.py
@@ -12,7 +12,7 @@ import pytest
 import yaml
 from pytest_operator.plugin import OpsTest
 
-from ..helpers import CHARM_SERIES, assert_sync_standbys, scale_application
+from ..helpers import CHARM_SERIES, assert_sync_standbys, get_leader_unit, scale_application
 from ..juju_ import juju_major_version
 from .helpers import (
     build_connection_string,
@@ -362,9 +362,16 @@ async def test_relation_data_is_updated_correctly_when_scaling(ops_test: OpsTest
         )
 
         # Remove the original units.
-        await ops_test.model.applications[DATABASE_APP_NAME].destroy_units(*units_to_remove)
+        leader_unit = await get_leader_unit(ops_test, DATABASE_APP_NAME)
+        await ops_test.model.applications[DATABASE_APP_NAME].destroy_units(*[
+            unit for unit in units_to_remove if unit != leader_unit.name
+        ])
         await ops_test.model.wait_for_idle(
-            apps=[DATABASE_APP_NAME], status="active", timeout=1500, wait_for_exact_units=2
+            apps=[DATABASE_APP_NAME], status="active", timeout=600, wait_for_exact_units=3
+        )
+        await ops_test.model.applications[DATABASE_APP_NAME].destroy_units(leader_unit.name)
+        await ops_test.model.wait_for_idle(
+            apps=[DATABASE_APP_NAME], status="active", timeout=600, wait_for_exact_units=2
         )
 
         # Get the updated connection data and assert it can be used
@@ -569,6 +576,33 @@ async def test_invalid_extra_user_roles(ops_test: OpsTest):
         )
         await ops_test.model.wait_for_idle(
             apps=[DATABASE_APP_NAME],
+            status="active",
+            raise_on_blocked=False,
+            timeout=1000,
+        )
+
+
+@pytest.mark.group(1)
+async def test_nextcloud_db_blocked(ops_test: OpsTest, charm: str) -> None:
+    async with ops_test.fast_forward():
+        # Deploy Nextcloud.
+        await ops_test.model.deploy(
+            "nextcloud",
+            channel="edge",
+            application_name="nextcloud",
+            num_units=1,
+        )
+        await ops_test.model.wait_for_idle(
+            apps=["nextcloud"],
+            status="blocked",
+            raise_on_blocked=False,
+            timeout=1000,
+        )
+
+        await ops_test.model.relate("nextcloud:database", f"{DATABASE_APP_NAME}:database")
+
+        await ops_test.model.wait_for_idle(
+            apps=[DATABASE_APP_NAME, "nextcloud"],
             status="active",
             raise_on_blocked=False,
             timeout=1000,
